@@ -2,13 +2,24 @@
   "name": "cache-bridge",
   "version": "0.0.4",
   "description": "",
-  "main": "index.js",
+  "main": "./index.js",
+  "types": "./index.d.ts",
   "files": [
-    "/lib"
+    "./**/*.js",
+    "./**/*.ts"
   ],
   "scripts": {
-    "test": "mocha --recursive ./test/test",
-    "lint": "eslint --max-warnings 0 ."
+    "build:rm": "rm -rf ./dist",
+    "build:tsc": "tsc --project tsconfig.build.json",
+    "build:copy": "mkdir -p dist && cp package.json package-lock.json README.md LICENSE dist/",
+    "build": "npm run build:rm && npm run build:tsc && npm run build:copy",
+    "publish:beta": "cd dist && npm publish --tag beta",
+    "publish:main": "cd dist && npm publish",
+    "test:ts": "mocha --require ts-node/register test/test/**/*.test.ts",
+    "test": "npm run test:ts",
+    "lint:ts": "tsc --noEmit",
+    "lint:eslint": "eslint --max-warnings 0 .",
+    "lint": "npm run lint:ts && npm run lint:eslint"
   },
   "repository": {
     "type": "git",
@@ -20,19 +31,27 @@
   "author": "SoftwareSing",
   "license": "MIT",
   "dependencies": {
-    "uuid": "^8.3.2"
+    "uuid": "^9.0.0"
   },
   "devDependencies": {
-    "@faker-js/faker": "^7.3.0",
-    "@types/chai": "^4.3.1",
-    "@types/mocha": "^9.1.1",
-    "chai": "^4.3.6",
-    "eslint": "^8.21.0",
-    "eslint-config-standard": "^17.0.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-n": "^15.2.4",
-    "eslint-plugin-promise": "^6.0.0",
-    "lru-cache": "^7.13.2",
-    "mocha": "^10.0.0"
+    "@faker-js/faker": "^7.6.0",
+    "@tsconfig/node14": "^1.0.3",
+    "@types/chai": "^4.3.5",
+    "@types/mocha": "^10.0.1",
+    "@types/node": "^18.16.5",
+    "@types/uuid": "^9.0.1",
+    "@typescript-eslint/eslint-plugin": "^5.59.2",
+    "chai": "^4.3.7",
+    "eslint": "^8.40.0",
+    "eslint-config-standard-with-typescript": "^34.0.1",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-n": "^15.7.0",
+    "eslint-plugin-promise": "^6.1.1",
+    "ioredis": "^5.3.2",
+    "lru-cache": "^9.1.1",
+    "mocha": "^10.2.0",
+    "redis": "^4.6.6",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.0.4"
   }
 }
