@@ -1,19 +1,25 @@
-const { faker } = require('@faker-js/faker')
+import { faker } from '@faker-js/faker'
 
-const typeList = ['undefined', 'null', 'number', 'boolean', 'string', 'object', 'array']
+interface TypeEnable {
+  undefined?: boolean
+  null?: boolean
+  number?: boolean
+  boolean?: boolean
+  string?: boolean
+  object?: boolean
+  array?: boolean
+}
+const typeList: Array<keyof TypeEnable> = ['undefined', 'null', 'number', 'boolean', 'string', 'object', 'array']
 
-/**
- * @param {{ undefined: Boolean, null: Boolean, number: Boolean, boolean: Boolean, string: Boolean, object: Boolean, array: Boolean }} [typeEnable]
- */
-exports.getRandomData = function getRandomData (typeEnable = {}) {
-  let defaultEnable = !Object.values(typeEnable)[0]
-  if (typeof defaultEnable !== 'boolean') defaultEnable = true
+export function getRandomData (typeEnable: TypeEnable = {}): any {
+  let defaultEnable = Boolean(Object.values(typeEnable)[0])
+  defaultEnable = !defaultEnable
 
   const availableTypeList = typeList.filter((type) => {
     const enable = typeEnable[type]
     return typeof enable === 'boolean' ? enable : defaultEnable
   })
-  if (availableTypeList.length < 1) return undefined
+  if (availableTypeList.length < 1) throw new Error('At least one type needs to be allowed')
 
   const type = faker.helpers.arrayElement(availableTypeList)
   switch (type) {
@@ -27,8 +33,8 @@ exports.getRandomData = function getRandomData (typeEnable = {}) {
   }
 }
 
-function getRandomObject () {
-  const result = {}
+function getRandomObject (): object {
+  const result: Record<string, any> = {}
   const keyNum = faker.datatype.number(20)
   for (let i = 0; i < keyNum; i += 1) {
     result[faker.random.alphaNumeric(i + 1)] = getRandomBaseData()
@@ -36,13 +42,13 @@ function getRandomObject () {
   return result
 }
 
-function getRandomArray () {
+function getRandomArray (): Array<string | number | boolean> {
   return (new Array(faker.datatype.number(20)))
-    .fill()
+    .fill(undefined)
     .map(getRandomBaseData)
 }
 
-function getRandomBaseData () {
+function getRandomBaseData (): string | number | boolean {
   const getFunc = faker.helpers.arrayElement([
     faker.datatype.number,
     faker.datatype.boolean,
