@@ -1,15 +1,11 @@
 module.exports = {
-  env: {
-    commonjs: true,
-    es2021: true,
-    node: true
-  },
   extends: [
-    'standard'
+    'standard-with-typescript'
   ],
   parserOptions: {
-    ecmaVersion: 'latest'
+    project: './tsconfig.json'
   },
+  plugins: [],
   rules: {
   }
 }
