@@ -1,8 +1,4 @@
-/**
- * @template T
- * @param {Iterable<T>} list
- */
-exports.shuffle = function shuffle (list) {
+export function shuffle <T> (list: Iterable<T>): T[] {
   const result = [...list]
   for (let i = result.length - 1; i > 0; i -= 1) {
     const j = Math.floor(Math.random() * (i + 1))
